D-STAR->FREE STAR*, Better represent ircddb status
Technically D-STAR is copyrighted by ICOM even though JARL created it
so we best get folks used to using the FREE STAR* name. Changed the
report on ircddb.net so that if the  gwys.txt file is broken but
ircddb.net is not, represent ircddb.net as ok until we have something
to compare it against.